removed deprecated XmlParser
